Remove -S flag from flawfinder params
The -S flag makes flawfinder outputs one warning per line. While this
seems to be the desired behavior, since most of the other tools used
here use the same approach, this flag is not compatible with firehose,
which we may use to have all the results in the same format.